add search components
