+++ mavis/test/pages/schools/schools_search_page.py
@@ -3,19 +3,18 @@ from playwright.sync_api import Page, expect
 from mavis.test.annotations import step
 from mavis.test.data_models import School
 from mavis.test.pages.header_component import HeaderComponent
+from mavis.test.pages.search_components import BaseSearchComponent
 
 
 class SchoolsSearchPage:
     def __init__(self, page: Page) -> None:
         self.page = page
+        self.search = BaseSearchComponent(page)
         self.header = HeaderComponent(page)
-        self.search_textbox = page.get_by_role("searchbox", name="Search")
-        self.search_button = page.get_by_role("button", name="Search")
 
     @step("Click on {1}")
     def click_school(self, school: School) -> None:
-        self.search_textbox.fill(str(school))
-        self.search_button.click()
+        self.search.search_for(str(school))
 
         self.page.get_by_role("link", name=str(school)).first.click()
