+++ mavis/test/pages/search_components/__init__.py
@@ -1,0 +1,9 @@
+from .base_search_component import BaseSearchComponent
+from .patient_search_component import PatientSearchComponent
+from .patient_status_search_component import PatientStatusSearchComponent
+
+__all__ = [
+    "BaseSearchComponent",
+    "PatientSearchComponent",
+    "PatientStatusSearchComponent",
+]
