+++ mavis/test/pages/sessions/sessions_children_page.py
@@ -6,7 +6,7 @@ from mavis.test.constants import (
 )
 from mavis.test.data_models import Child
 from mavis.test.pages.header_component import HeaderComponent
-from mavis.test.pages.sessions.search_component import SearchComponent
+from mavis.test.pages.search_components import PatientStatusSearchComponent
 from mavis.test.pages.sessions.sessions_tabs import SessionsTabs
 from mavis.test.utils import (
     reload_until_element_is_visible,
@@ -17,29 +17,9 @@ class SessionsChildrenPage:
     def __init__(self, page: Page) -> None:
         self.page = page
         self.tabs = SessionsTabs(page)
-        self.search = SearchComponent(page)
+        self.search = PatientStatusSearchComponent(page)
         self.header = HeaderComponent(page)
 
-        self.needs_consent_radio = self.page.get_by_role(
-            "radio",
-            name="Needs consent",
-        )
-        self.has_a_refusal_radio = self.page.get_by_role(
-            "radio",
-            name="Has a refusal",
-        )
-        self.parent_refused_checkbox = self.page.get_by_role(
-            "checkbox",
-            name="Parent refused",
-        )
-        self.due_vaccination_radio = self.page.get_by_role(
-            "radio",
-            name="Due vaccination",
-        )
-        self.conflicting_consent_checkbox = self.page.get_by_role(
-            "checkbox",
-            name="Conflicting consent",
-        )
         self.attending_button = self.page.get_by_role("button", name="Attending").first
 
     def register_child_as_attending(self, child: Child) -> None:
@@ -52,26 +32,6 @@ class SessionsChildrenPage:
     @step("Click on Attending")
     def click_on_attending(self) -> None:
         self.attending_button.click()
-
-    @step("Select Needs consent")
-    def select_needs_consent(self) -> None:
-        self.needs_consent_radio.check()
-
-    @step("Select Has a refusal")
-    def select_has_a_refusal(self) -> None:
-        self.has_a_refusal_radio.check()
-
-    @step("Select Parent refused")
-    def select_parent_refused(self) -> None:
-        self.parent_refused_checkbox.check()
-
-    @step("Select Due vaccination")
-    def select_due_vaccination(self) -> None:
-        self.due_vaccination_radio.check()
-
-    @step("Select Conflicting consent")
-    def select_conflicting_consent(self) -> None:
-        self.conflicting_consent_checkbox.check()
 
     def verify_child_shows_correct_flu_consent_method(
         self,
@@ -95,6 +55,8 @@ class SessionsChildrenPage:
 
         return flu_consent_section
 
-    @step("Expect Has a refusal to be selected")
-    def expect_has_a_refusal_to_be_selected(self) -> None:
-        expect(self.has_a_refusal_radio).to_be_checked()
+    @step("Check note {2} appears in search for {1}")
+    def check_note_appears_in_search(self, child: Child, note: str) -> None:
+        heading = self.page.get_by_role("heading", name=str(child))
+        next_element = heading.locator("xpath=following-sibling::*[1]")
+        expect(next_element.get_by_role("blockquote")).to_have_text(note)
