+++ mavis/test/pages/sessions/sessions_record_vaccinations_page.py
@@ -1,7 +1,7 @@
 from playwright.sync_api import Page
 
 from mavis.test.pages.header_component import HeaderComponent
-from mavis.test.pages.sessions.search_component import SearchComponent
+from mavis.test.pages.search_components import PatientSearchComponent
 from mavis.test.pages.sessions.sessions_tabs import SessionsTabs
 
 
@@ -9,5 +9,5 @@ class SessionsRecordVaccinationsPage:
     def __init__(self, page: Page) -> None:
         self.page = page
         self.tabs = SessionsTabs(page)
-        self.search = SearchComponent(page)
+        self.search = PatientSearchComponent(page)
         self.header = HeaderComponent(page)
