+++ tests/test_import_offline_vaccinations.py
@@ -235,10 +235,10 @@ def test_vaccination_file_upload_creates_child_no_setting(
     ImportsPage(page).header.click_mavis_header()
     DashboardPage(page).click_children()
 
-    ChildrenSearchPage(page).click_advanced_filters()
-    ChildrenSearchPage(page).check_children_aged_out_of_programmes()
-    ChildrenSearchPage(page).search_with_all_filters_for_child_name(str(child))
-    ChildrenSearchPage(page).click_record_for_child(child)
+    ChildrenSearchPage(page).search.click_advanced_filters()
+    ChildrenSearchPage(page).search.check_children_aged_out_of_programmes()
+    ChildrenSearchPage(page).search.search_for_child_name_with_all_filters(str(child))
+    ChildrenSearchPage(page).search.click_child(child)
     ChildRecordPage(page).click_vaccination_details(school)
     VaccinationRecordPage(page).expect_vaccination_details("Location", str(school))
