+++ tests/test_online_consent_school_moves.py
@@ -117,7 +117,7 @@ def test_online_consent_school_moves_with_existing_patient(
     )
 
     SessionsOverviewPage(page).tabs.click_children_tab()
-    SessionsChildrenPage(page).select_due_vaccination()
+    SessionsChildrenPage(page).search.select_due_vaccination()
     SessionsChildrenPage(page).search.search_for(str(child))
     SessionsChildrenPage(page).verify_child_shows_correct_flu_consent_method(
         child, ConsentOption.NASAL_SPRAY_OR_INJECTION
@@ -184,7 +184,7 @@ def test_online_consent_school_moves_with_new_patient(
     )
 
     SessionsOverviewPage(page).tabs.click_children_tab()
-    SessionsChildrenPage(page).select_due_vaccination()
+    SessionsChildrenPage(page).search.select_due_vaccination()
     SessionsChildrenPage(page).search.search_for(str(child))
     SessionsChildrenPage(page).verify_child_shows_correct_flu_consent_method(
         child, ConsentOption.NASAL_SPRAY_OR_INJECTION
